set signup route & styled signup form
