+++ public/styles/main-style.css
@@ -302,20 +302,24 @@ nav img#avatar {
     margin-bottom: 21px;
 }
 
-/*Login Page*/
-#login {
+/*Login & Sign Up Page*/
+#login,
+#signup {
     display: flex;
     flex-direction: column;
     align-items: center;
 }
 
 #login .card,
-#login .error {
+#login .error,
+#signup .card,
+#signup .error  {
     width: 320px;
     margin-top: 36px;
 }
 
-#login .card .big-text {
+#login .card .big-text,
+#signup .card .big-text  {
     margin: 36px 0;
 }
