--- views/register.hbs
@@ -1,37 +1,0 @@
-<h2>Register</h2>
-<hr style="margin-top: 5px;">
-
-{{#if success}}
-<div class="alert alert-success">
-    <strong>{{success}}</strong>
-</div><a class="btn btn-success pull-right" href="/login">Proceed to Login</a>
-<br><br><br>
-{{else}}
-{{#if error}}
-<div class="alert alert-danger"> <strong> {{error}} </strong></div>
-{{/if}}
-<form name="register" method="post" action="/register">
-    <div class="form-row">
-        <div class="form-group col-md-12" style="padding-left: 0;">
-            <input type="text" class="form-control" id="username" name="username" placeholder="Username" required>
-        </div>
-    </div>
-
-    <div class="form-row">
-        <div class="form-group col-md-12" style="padding-left: 0;">
-            <input type="password" class="form-control" id="password" name="password" placeholder="Password" required>
-        </div>
-    </div>
-
-    <div class="form-row">
-        <div class="form-group col-md-12" style="padding-left: 0;">
-            <input type="password" class="form-control" id="confirm_password" name="confirm_password"
-                placeholder="Confirm Password" required>
-        </div>
-    </div>
-
-    <div class="form-row col-md-12 text-right" style="padding-left: 0;">
-        <button type="submit" class="btn btn-primary btn-success">Register</button>
-    </div>
-</form>
-{{/if}}
