+++ views/signup.hbs
@@ -1,12 +1,12 @@
-<div id="login">
+<div id="signup">
 
     {{#if error}}
     <div class="error"> {{error}}</div>
     {{/if}}
 
     <div class="card">
-        <h2 class="big-text">Welcome back!</h2>
-        <form name="login" method="post" action="/login">
+        <h2 class="big-text">Join Us!</h2>
+        <form name="signup" method="post" action="/signup">
             <div>
                 <p class="label">Username</p>
                 <input type="text" class="form-control" id="username" name="username" required>
@@ -15,8 +15,12 @@
                 <p class="label">Password</p>
                 <input type="password" class="form-control" id="password" name="password" required>
             </div>
-            <button>Login</button>
+            <div>
+                <p class="label">Confirm Password</p>
+                 <input type="password" class="form-control" id="confirm_password" name="confirm_password" required>
+            </div>
+            <button>Create my account</button>
         </form>
     </div>
-    
+
 </div>
